Organise tests to match package structure
