+++ tests/test_diags/test_lifecycle.py
@@ -1,0 +1,15 @@
+import huracanpy
+
+
+def test_time_from_genesis():
+    data = huracanpy.load(huracanpy.example_csv_file, tracker="csv")
+    t = huracanpy.diags.lifecycle.time_from_genesis(data)
+    assert t.min() == 0
+    assert t.max() == 1166400000000000
+
+
+def test_time_from_extremum():
+    data = huracanpy.load(huracanpy.example_csv_file, tracker="csv")
+    t = huracanpy.diags.lifecycle.time_from_extremum(data, "wind10", "max")
+    assert t.min() == -972000000000000
+    assert t.max() == 367200000000000
